modify category module

+++ src/app/modules/Brand/brand.service.ts
@@ -1,5 +1,6 @@
 import { ErrorType } from "../../constant/global.constant";
 import AppError from "../../errors/AppError";
+import Product from "../Product/product.model";
 import { TBrand } from "./brand.interface";
 import Brand from "./brand.model";
 
@@ -48,6 +49,14 @@ const updateBrandIntoDB = async (brandId: string, data: Partial<TBrand>) => {
 const deleteBrandIntoDB = async (brandId: string) => {
   if (!(await Brand.findById(brandId))) {
     throw new AppError(400, `invalid brand id`, ErrorType.validation);
+  }
+
+  if (await Product.findOne({ brand: brandId })) {
+    throw new AppError(
+      400,
+      `This brand already used, can't delete it`,
+      ErrorType.badRequest,
+    );
   }
 
   await Brand.findByIdAndDelete(brandId);
